Code refactored. New Readme. New Contributing doc.
