+++ chibi/services/providers/tools/file_editor.py
@@ -6,13 +6,14 @@ This module provides functions for efficiently editing files through LLM tools.
 import os
 import re
 from pathlib import Path
-from typing import Unpack
+from typing import Any, Unpack
 
 from loguru import logger
 from openai.types.chat import ChatCompletionToolParam
 from openai.types.shared_params import FunctionDefinition
 
 from chibi.config import gpt_settings
+from chibi.services.providers.tools.exceptions import ToolException
 from chibi.services.providers.tools.tool import ChibiTool
 from chibi.services.providers.tools.utils import AdditionalOptions
 
@@ -810,3 +811,65 @@ class InsertBeforePatternTool(ChibiTool):
         except Exception as e:
             logger.error(f"[{kwargs.get('model', 'Unknown model')}] Error inserting before pattern in {full_path}: {e}")
             raise
+
+
+class CreateFileTool(ChibiTool):
+    register = gpt_settings.filesystem_access
+    definition = ChatCompletionToolParam(
+        type="function",
+        function=FunctionDefinition(
+            name="create_file",
+            description="Create a file at the given full path (including any directories).",
+            parameters={
+                "type": "object",
+                "properties": {
+                    "content": {
+                        "type": "string",
+                        "description": "File content",
+                    },
+                    "full_path": {
+                        "type": "string",
+                        "description": "Full file name, including file path.",
+                    },
+                    "overwrite": {
+                        "type": "string",
+                        "enum": ["true", "false"],
+                        "description": (
+                            "Set it to true to overwrite the file. If overwrite is false and "
+                            "the file exists, raises FileExistsError."
+                        ),
+                    },
+                },
+                "required": ["content", "full_path", "overwrite"],
+            },
+        ),
+    )
+    name = "create_file"
+
+    @classmethod
+    async def function(
+        cls,
+        full_path: str,
+        content: str,
+        overwrite: str = "false",
+        encoding: str = "utf-8",
+        **kwargs: Unpack[AdditionalOptions],
+    ) -> dict[str, Any]:
+        try:
+            path = Path(full_path).expanduser().resolve()
+            parent = path.parent
+            parent.mkdir(parents=True, exist_ok=True)
+
+            if path.exists():
+                if overwrite != "true":
+                    raise FileExistsError(f"File {path} already exists")
+                logger.log("TOOL", f"File {path} exists. Overwriting.")
+
+            with path.open("w", encoding=encoding) as f:
+                f.write(content)
+
+            logger.log("TOOL", f"File {path} created successfully.")
+            return {"file": str(path)}
+
+        except Exception as e:
+            raise ToolException(f"Failed to create file {full_path}. Error: {e}")
