+++ chibi/services/user.py
@@ -100,6 +100,7 @@ async def get_llm_chat_completion_answer(
         prompt = {
             "prompt": user_message,
             "datetime_now": datetime.datetime.now(timezone.utc).strftime("%Y-%m-%d %H:%M:%S %Z%z"),
+            "type": "user message",
             "transcribed_from_voice_message": bool(user_voice_message),
         }
 
@@ -208,6 +209,22 @@ async def get_info(db: Database, user_id: int) -> str:
 async def set_info(db: Database, user_id: int, new_info: str) -> None:
     user = await db.get_or_create_user(user_id=user_id)
     user.info = new_info
+    await db.save_user(user)
+
+
+@inject_database
+async def activate_llm_skill(db: Database, user_id: int, skill_name: str, skill_payload: str) -> None:
+    user = await db.get_or_create_user(user_id=user_id)
+    user.llm_skills[skill_name] = skill_payload
+    await db.save_user(user)
+
+
+@inject_database
+async def deactivate_llm_skill(db: Database, user_id: int, skill_name: str) -> None:
+    user = await db.get_or_create_user(user_id=user_id)
+    if skill_name not in user.llm_skills.keys():
+        raise ValueError(f"The skill {skill_name} seems never been activated")
+    user.llm_skills.pop(skill_name)
     await db.save_user(user)
